package/pyproject.toml
[build-system]

requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

["project"]
name = "arquivo_pt"
dynamic = ["version"] 
description = "Package to Support 25 of April Arquivo PT submission"
readme = "README.md"

authors = [
    {name = "Rúben Almeida", email = "ruben.f.almeida@inesctec.pt"},
    {name = "Manuel Coutinho"}
]
requires-python = ">=3.10"

dependencies = [
    "clean-text[gpl]>=0.6.0",
    "tqdm>=4.66.5",
    "requests>=2.32.3",
    "json-repair>=0.23.1",
    "environs>=11.0.0",
    "numpy",
    "pandas>=2.2.3",
    "beautifulsoup4>=4.13.4",
]

[project.optional-dependencies]

test = [
    "pytest>=8.3.3",
    "levenshtein",
]

llm = [
    "google-genai>=1.10.0",
    "openai>=1.73.0",
]

langchain = [
    "langchain>=0.3.16",
    "langchain-community>=0.3.16",
    "nltk>=3.9.1",
]

demo = [
    "fastapi[standard]>=0.115.2",
    "python-multipart>=0.0.20",
    "arquivo_pt[langchain]",
    "arquivo_pt[llm]",
    "cryptography>=44.0.2",
]

dev = [    
    "arquivo_pt[demo]",
    "arquivo_pt[test]",
    "arquivo_pt[langchain]",
    "arquivo_pt[llm]"
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["arquivo_pt", "arquivo_pt.*"]
namespaces = false

[tool.setuptools_scm]
